refactor(apple): organize app sources and share workspace coordination
