+++ apps/apple/Packages/SokosumiWorkspace/.gitignore
@@ -1,0 +1,2 @@
+.build/
+.swiftpm/
