+++ apps/apple/Packages/SokosumiWorkspace/Tests/SokosumiWorkspaceTests/StubOAuthBrowser.swift
@@ -1,0 +1,11 @@
+import Foundation
+import SokosumiAuth
+
+@MainActor
+final class StubOAuthBrowser: OAuthBrowser {
+  func authenticate(url _: URL) async throws -> URL {
+    throw OAuthBrowserError.couldNotStart
+  }
+
+  func cancel() {}
+}
