+++ apps/apple/Sokosumi/App/CoreSettings.swift
