+++ apps/apple/Sokosumi/Authentication/MacOAuthBrowser.swift
