+++ apps/apple/Sokosumi/Chat/Rendering/CoworkerThoughtView.swift
