+++ apps/apple/Sokosumi/Settings/SettingsView.swift
@@ -1,4 +1,5 @@
 import SokosumiAuth
+import SokosumiWorkspace
 import SwiftUI
 
 /// Minimal Settings: account identity plus the Core endpoint in use.
