+++ apps/apple/SokosumiTests/App/CoreSettingsTests.swift
