+++ apps/apple/SokosumiTests/Chat/Rendering/NativeHighlightResourceTests.swift
